backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
wfdb==4.1.2
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
